show data home,profile, add tagservice, update post-api-image,(error can't decide whether to keep or delete the image when editing a post

+++ src/main/java/net/blwsmartware/controller/home/post/PostController.java
@@ -12,9 +12,14 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.Part;
 import net.blwsmartware.constant.IConstant;
+import net.blwsmartware.dao.IPostHasTagDAO;
+import net.blwsmartware.model.PostHasTagModel;
 import net.blwsmartware.model.PostModel;
+import net.blwsmartware.model.TagModel;
 import net.blwsmartware.model.UserModel;
+import net.blwsmartware.service.IPostHasTagService;
 import net.blwsmartware.service.IPostService;
+import net.blwsmartware.service.ITagService;
 import net.blwsmartware.service.IUserService;
 import net.blwsmartware.service.impl.ImageServiceImpl;
 import net.blwsmartware.util.FormUtil;
@@ -25,6 +30,7 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStreamReader;
+import java.text.Normalizer;
 import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -47,6 +53,12 @@ import java.util.stream.Collectors;
         @Inject
         private ImageServiceImpl imageService;
 
+        @Inject
+        private ITagService tagService;
+
+        @Inject
+        private IPostHasTagService postHasTagService;
+
         protected void doGet(HttpServletRequest request, HttpServletResponse response)
                 throws ServletException, IOException {
             String token = JWTUtil.getToken(request);
@@ -58,6 +70,8 @@ import java.util.stream.Collectors;
             if(model.getId()!=null){
                 model = postService.findByID(model.getId());
             }
+            UserModel userModel = getUserInfo(request);
+
             request.setAttribute("postModel", model);
             RequestDispatcher rd = request.getRequestDispatcher("/views/web/create-post.jsp");
             rd.forward(request, response);
@@ -65,71 +79,30 @@ import java.util.stream.Collectors;
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-            String method = request.getParameter("_method");
-            if(method.equalsIgnoreCase("put")){
-                doPut(request, response);
-                return;
-            }
         request.setCharacterEncoding("UTF-8");
         response.setContentType("application/json");
-        Map<String, Object> formData ;
+        Map<String, Object> formData = new HashMap<>();
+        Map<String, Object> result = new HashMap<>();
+        ObjectMapper objectMapper = new ObjectMapper();
+        File rez = null;
         try {
-            Long idUser = JWTUtil.getIdUser(request);
-            UserModel userModel = userService.findByID(idUser);
-            formData = new HashMap<>();
-            File rez = null;
-            for (Part part : request.getParts()) {
-                if (part.getContentType() == null) {
-                    String fieldName = part.getName();
-                    String fieldValue = new BufferedReader(new InputStreamReader(part.getInputStream()))
-                            .lines().collect(Collectors.joining("\n"));
-                    formData.put(fieldName, fieldValue);
-                } else {
-                    if (part.getName().equals("image")) {
-                        try {
-                            imageService.makeDir();
-                            imageService.setPath(part);
-                            imageService.saveImageToDisk();
-                            imageService.createThumbnail();
-                            rez = imageService.resize();
-                        } catch (Exception e) {
-                            System.out.println(e.getMessage());
-                        }
-                    }
-                }
-            }
+            UserModel userModel = getUserInfo(request);
+            PostModel postModel = getPostInfo(request);
 
-
-            PostModel postModel = new ObjectMapper().convertValue(formData, PostModel.class);
-
-
-            String tagList = request.getParameter("tag-collect");
-            Pattern pattern = Pattern.compile("#\\w+");
-            Matcher matcher = pattern.matcher(tagList);
-            List<String> tags = new ArrayList<>();
-            while (matcher.find()){
-                tags.add(matcher.group());
-            }
-
-            if(!tags.isEmpty()){
-                postModel.setTab(tags);
-            }
-
-            if (rez != null) {
-                postModel.setImagePath(rez.getPath());
-                postModel.setThumbnail(imageService.getThumbnailName());
-            }
             postModel.setCreatedBy(userModel.getName());
             postModel.setAuthId(userModel.getId());
             postModel = postService.save(postModel);
-            System.out.println(postModel);
+            saveTag(request,userModel, postModel);
 
-            response.sendRedirect(request.getContextPath() + "/profile");
+            System.out.println(postModel);
+            result.put("status","success");
         } catch (Exception e) {
-            request.setAttribute("error-post", "error");
-            request.getRequestDispatcher("/views/web/create-post.jsp").forward(request, response);
+            result.put("status", "error");
+            result.put("message", e.getMessage());
         }
-    }
+        objectMapper.writeValue(response.getOutputStream(), result);
+
+}
 
     @Override
     protected void doPut(HttpServletRequest request, HttpServletResponse response)
@@ -138,13 +111,44 @@ import java.util.stream.Collectors;
         response.setContentType("application/json");
 
         Map<String, Object> formData;
+        Map<String, Object> result = new HashMap<>();
+        ObjectMapper objectMapper = new ObjectMapper();
+
         try {
+            UserModel userModel = getUserInfo(request);
+            PostModel postModel = getPostInfo(request);
+
+            postModel.setModifiedBy(userModel.getName());
+            postModel = postService.update(postModel);
+            saveTag(request,userModel, postModel);
+            System.out.println(postModel);
+            result.put("status", "success");
+
+        } catch (Exception e) {
+            result.put("status","error");
+            result.put("message", e.getMessage());
+        }
+        objectMapper.writeValue(response.getOutputStream(), result);
+
+    }
+    private List<String> getTags(HttpServletRequest request){
+        String tagList = request.getParameter("tag-collect");
+        Pattern pattern = Pattern.compile("#(\\S+)");
+        Matcher matcher = pattern.matcher(tagList);
+        List<String> tags = new ArrayList<>();
+        while (matcher.find()) {
+            tags.add(matcher.group(1));
+        }
+        return tags;
+    }
+    private UserModel getUserInfo(HttpServletRequest request){
             Long idUser = JWTUtil.getIdUser(request);
-            UserModel userModel = userService.findByID(idUser);
-            formData = new HashMap<>();
-            File rez = null;
-
-
+            return userService.findByID(idUser);
+    }
+    private PostModel getPostInfo(HttpServletRequest request){
+        Map<String, Object> formData = new HashMap<>();
+        File rez = null;
+        try {
             for (Part part : request.getParts()) {
                 if (part.getContentType() == null) {
                     String fieldName = part.getName();
@@ -159,46 +163,58 @@ import java.util.stream.Collectors;
                             imageService.saveImageToDisk();
                             imageService.createThumbnail();
                             rez = imageService.resize();
+
                         } catch (Exception e) {
                             System.out.println(e.getMessage());
                         }
                     }
                 }
             }
-
-            PostModel postModel = new ObjectMapper().convertValue(formData, PostModel.class);
+        }catch (Exception e){
+            System.out.println(e.getMessage());
+        }
+        PostModel postModel =  new ObjectMapper().convertValue(formData, PostModel.class);
+        if(postModel.getId()!=null){
             postModel = postService.findByID(postModel.getId());
-            if(!imageService.delete(postModel.getImagePath())){
-                System.out.println("Delete image unsuccessfully");
+//            String imageAction = request.getParameter("imageAction");
+//            if(imageAction.equalsIgnoreCase("delete")){
+            if(rez != null){
+                if(!imageService.delete(postModel.getImagePath())){
+                    System.out.println("Delete image unsuccessfully");
+                }
             }
+//            }
+            postService.deleteAllTag(postModel);
 
-            String tagList = request.getParameter("tag-collect");
-            Pattern pattern = Pattern.compile("#\\w+");
-            Matcher matcher = pattern.matcher(tagList);
-            List<String> tags = new ArrayList<>();
-            while (matcher.find()) {
-                tags.add(matcher.group());
-            }
-            //Chưa xóa tag đã đăng trước đó nếu có thay đổi
-            if (!tags.isEmpty()) {
-                postModel.setTab(tags);
-            }
-
-            if (rez != null) {
-                String pathImage = imageService.getPathToSaveDB(rez.getPath());
-                postModel.setImagePath(pathImage);
-                postModel.setThumbnail(imageService.getThumbnailName());
-            }
-            postModel.setModifiedBy(userModel.getName());
-            postModel = postService.update(postModel);
-            System.out.println(postModel);
-            response.sendRedirect(request.getContextPath() + "/profile");
-
-        } catch (Exception e) {
-            request.setAttribute("error-post", "error");
-            request.getRequestDispatcher("/views/web/create-post.jsp").forward(request, response);
+        }
+        List<String> tags = getTags(request);
+        if (!tags.isEmpty()) {
+            postModel.setTab(tags);
         }
 
+        if (rez != null) {
+            postModel.setImagePath(rez.getName());
+            postModel.setThumbnail(imageService.getThumbnailName());
+        }
+        return postModel;
+    }
+    private void saveTag(HttpServletRequest request, UserModel userModel, PostModel postModel){
+            List<String> tags = getTags(request);
+        if(!tags.isEmpty()){
+            for(String tag : tags){
+                TagModel tagModel = tagService.findByTagName(tag);
+                if(tagModel == null ||tagModel.getName() == null){
+                    tagModel = new TagModel();
+                    tagModel.setCreatedBy(userModel.getName());
+                    tagModel.setName(tag);
+                    tagModel = tagService.save(tagModel);
+                }
+                PostHasTagModel postHasTagModel = new PostHasTagModel();
+                postHasTagModel.setTagId(tagModel.getId());
+                postHasTagModel.setPostId(postModel.getId());
+                postHasTagService.save(postHasTagModel);
+            }
+        }
     }
